BC Break - move adapter v0.8
